Define CpuDeviceInterface (#636)
